requirements.txt
streamlit
pandas
matplotlib
wordcloud
python-dotenv
requests
tqdm
spacy
beautifulsoup4
markdown
nltk
spacy
google.generativeai